pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "streamlit-authenticator>=0.4.2",
    "streamlit>=1.44.1",
    "pyyaml>=6.0.2",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.40",
    "pyinstaller>=6.12.0",
    "twilio>=9.5.2",
    "numpy>=2.2.4",
    "openpyxl>=3.1.5",
    "xlsxwriter>=3.2.2",
    "plotly>=6.0.1",
    "pytesseract>=0.3.13",
    "pdf2image>=1.17.0",
    "pillow>=11.1.0",
    "scikit-learn>=1.6.1",
    "statsmodels>=0.14.4",
    "matplotlib>=3.10.1",
]
